Rename security file
